refactor(token): move the directory
