+++ extractor/image/token/ecr/ecr.go
