+++ extractor/image/token/ecr/ecr_test.go
