+++ extractor/image/token/gcr/gcr.go
