+++ extractor/image/token/gcr/gcr_test.go
